requirements.txt
openai
python-dotenv
openai-agents